change(either): remove ComputeFunction export
for a more typescript / javascript oriented
experience, rather than using the `this` context
of the passed ComputeFunction, we will instead
pass the available context / functionality as a
class that can be destructured. This allows for
a more intuitive experience and also reduces a lot
of redundant type information.

Fixes N/A